+++ testing/upload_validation.ts
@@ -189,14 +189,16 @@ export function validateSyncTableSchema(schema: any): ArraySchema<ObjectSchema<a
   );
 }
 
-function getNonUniqueElements<T>(items: T[]): T[] {
-  const set = new Set<T>();
+function getNonUniqueElements<T extends string>(items: T[]): T[] {
+  const set = new Set<string>();
   const nonUnique: T[] = [];
   for (const item of items) {
-    if (set.has(item)) {
+    // make this case insensitive
+    const normalized = item.toUpperCase()
+    if (set.has(normalized)) {
       nonUnique.push(item);
     }
-    set.add(item);
+    set.add(normalized);
   }
   return nonUnique;
 }
@@ -1202,8 +1204,34 @@ const unrefinedPackVersionMetadataSchema = zodCompleteObject<PackVersionMetadata
     message: 'Formula namespaces can only contain alphanumeric characters and underscores.',
   }),
   systemConnectionAuthentication: z.union(zodUnionInput(systemAuthenticationValidators)).optional(),
-  formulas: z.array(formulaMetadataSchema).max(Limits.BuildingBlockCountPerType).optional().default([]),
-  formats: z.array(formatMetadataSchema).max(Limits.BuildingBlockCountPerType).optional().default([]),
+  formulas: z
+    .array(formulaMetadataSchema)
+    .max(Limits.BuildingBlockCountPerType)
+    .optional()
+    .default([])
+    .superRefine((data, context) => {
+      const formulaNames = data.map(formulaDef => formulaDef.name);
+      for (const dupe of getNonUniqueElements(formulaNames)) {
+        context.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: `Formula names must be unique. Found duplicate name "${dupe}".`,
+        });
+      }
+    }),
+  formats: z
+    .array(formatMetadataSchema)
+    .max(Limits.BuildingBlockCountPerType)
+    .optional()
+    .default([])
+    .superRefine((data, context) => {
+      const formatNames = data.map(formatDef => formatDef.name);
+      for (const dupe of getNonUniqueElements(formatNames)) {
+        context.addIssue({
+          code: z.ZodIssueCode.custom,
+          message: `Format names must be unique. Found duplicate name "${dupe}".`,
+        });
+      }
+    }),
   syncTables: z
     .array(syncTableSchema)
     .max(Limits.BuildingBlockCountPerType)
@@ -1220,7 +1248,10 @@ const unrefinedPackVersionMetadataSchema = zodCompleteObject<PackVersionMetadata
             });
           }
         }
-        identityNames.push(tableDef.schema.identity?.name);
+        // only add identity names that are not undefined to check for dupes
+        if (tableDef.schema.identity) {
+          identityNames.push(tableDef.schema.identity?.name);
+        }
       }
       for (const dupe of getNonUniqueElements(identityNames)) {
         context.addIssue({